Refactored code to be DRY; click on list items results in strikethrough
